New upstream release;
0.53
The changes in this release are as follows:
Several noticeable performance improvements were made in this release.
A significant number of small bugs were fixed in several log formats.

0.52
The changes in this release are as follows:
Several minor bugs were fixed, including bugs in the energymech and
mIRC6 parsers. The options IgnoreWords, FoulWords, and ViolentWords
now support wildcards. Some minor layout changes to the generated
HTML page were made.