New upstream release;
0.5.12
The changes in this release are as follows:
There was an "off-by-one" error which resulted in a corrupted mirror
image.

0.5.11
The changes in this release are as follows:
In the previous release, "-O vts_01_4 -o -" didn't write to stdout,
but rather to the files. Info texts of vobcopy ended up in stdout
streams. These problems have been fixed. The new option -q (quiet)
has been added.