New upstream release, adding to 10.3, discontinuing the none ssl variant.
0.99.7
The changes in this release are as follows:
The memo preferences can now be hidden. The month and week views were
rewritten to be faster and resizable. An option to iconify at startup
was added. GTK2 support was expanded. cp1253 support was added. The
last categories are now remembered. Overdue todos are highlighted.
The possibility of memo category corruption was fixed. Many other
bugfixes and small improvements were made.