New upstream release;
0.57
The changes in this release are as follows:
This release fixed a problem with the NFiles option  when the
specified values was more than the actual  number of logfiles. A new
option for configuring the  number that triggers big numbers was
added. A small  warning was fixed.

0.58
The changes in this release are as follows:
A new ShowFoulDecimals feature for changing the  number of decimal
places was added. A bug with the  counting of reference nicks was
fixed, and the  documentation was cleaned up.

0.59
The changes in this release are as follows:
This release fixes a problem with the number of rows being displayed
in most active nicks when a nick was ignored. The Windows
documentation has been updated to hopefully cause less problems for
users. Some other minor enhancements were made.