Update Ruby to 1.8.3.
The Ruby tcltk and ssl extensions are packaged separately.  The ssl
extension since Fink has a crypto branch for ssl programs, and the
tcltk extension for linking against Fink's X11 version of tcltk, in
case Ruby needs to be linked against an Aqua version.

The Debian ruby1.8_1.8.3-2 package was used as a source for this
package.  Many of the same configure command line options are used.
Also, of the patches except for the 902_extra_search_path.patch from
that package are included in this package.  The three patches 00* are
not from the Debian package.