New upstream release
1.9.9
The changes in this release are as follows:
The code has been updated to work with OpenBabel 2.0. A few minor bugs
in file transfer and text labels were fixed. This release can export
directly to 3D modelling programs.

1.9.8
The changes in this release are as follows:
XDrawChem's database is now derived from the much larger PubChem
compound database. Minor fixes were made to super- and sub-script
drawing, font selection, screen painting, and symbol drawing.

1.9.7
The changes in this release are as follows:
A few bugs with screen drawing were fixed. A few new drawing symbols
were added. XDrawChem can now output InChI identifiers for
structures.

1.9.6
The changes in this release are as follows:
Several bugs with memory, drawing objects, and file I/O have been
fixed. The paint devices for screen and printer have been
consolidated, so print/image output should match what is on screen.
More options for text and bracket drawing have been added.