Revert to building libboost_python with Fink's python.
When compiled with the system python2.3 framework, the libboost_python.dylib was linked to the
framework python dylib, and this produced a crash when used with any other python.

We now build with any python from Fink, 2.3, 2.4, or 2.5.
The resulting libboost_python.dylib can be used with any other of these python versions.
(Fingers crossed).